NTWK-2550: fix libovndb cache update problem
This is to fix the cache update problem and remove the workaround of
sleeping after ovsdb transaction commit.

The problem is in depended github.com/cenkalti/rpc2 library, which
didn't preserve the order of incoming message handling.

To fix the library quickly, the dependencies are copied to vendor/
so that we can directly update.

The actual change to the rpc2 library is here:
https://github.com/cenkalti/rpc2/pull/13/commits/059c22f1643d92b5d86cd44a565c064a62d9e5de

Change-Id: I5aed3346aad5436d4db9f285e93a4a0fca2d8fba