Remove cenkalti from vendor and use upstream
cenkalti/rpc2 was copied to vendor director because it didn't
support blocking mode, so we had to copy and modify locally.
The change was critical to ensure correctness of transaction
commit. Now that our pull request has been merged [1], we can
remove it from the repo.

The blocking mode was added to upstream as an option, so in
socketplane/libovsdb we have to enable the option. The change
in socketplane/libovsdb is included in this patch. Unfortunately
the pull request [2] to socketplane/libovsdb is not merged yet
upstream, so we have to maintain socketplane/libovsdb under
/vendor in this repo.

[1] https://github.com/cenkalti/rpc2/pull/14

[2] https://github.com/socketplane/libovsdb/pull/58