Fix addresses format from slice to string
OVN NB expects the addresses to be a single string and not a slice.
Using a slice causes:
invalid syntax '10.128.2.3' in logical switch port addresses. No MAC address found
invalid syntax '10.128.2.3' in port security. No MAC address found

The GoSet object is a slice of addresses instead of a single string:
GoSet:[0a:58:0a:80:04:04 10.128.4.4]

Which causes a newline when we do ovn-nbctl lsp-get-addresses:

0a:58:0a:80:04:03
10.128.4.3

Signed-off-by: Tim Rozet <trozet@redhat.com>