add support for monitoring a subset of tables in OVN SB or NB DB
this commit adds a new field to Config struct called TableCols which
is of type map[string][]string (maps table names to columns). with this
user can specify what tables in NB or in SB they would like to Monitor.

in a large scale OVN K8s cluster, the SB DB will be huge with 1000s of
entries in various tables (Logical_Flow, Port_Binding, adn so on),
neither do we want to get the initial data for these tables nor we want
updates for these tables and this commit provides a way to specify the
tables the user is interested in

adding support for specific columns within a table needs some more work
since all of the rowTo<tableName>() functions need to be updated to
reflect missing fields in the Row (since client expressed interest in
only subset of rows)

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>