Fix LRSRAdd and LRSRDel output_port and policy handling.
Current implementation of these two APIs is broken when output_port
or policy is used.

Firstly, there is a typo in lrsrAddImp() for the column "policies"
which should be "policy".

Secondly, for the optional value it should use string in the row
structure instead of array. Using array end up as ovsdb server side
error complaining "syntax error ...".

This patch fixes both problems and added tests that cover both parameters.

Signed-off-by: Han Zhou <hzhou8@ebay.com>