test: Allow to set the source directory externally
Currently ovn and ovs are checked out in under `pwd`.
This is fine for a one-off build such as the one travis does. But for
for if we wanted to reuse the ovs build for many tests while keeping the
working directory clean, this is not enough

In order to support sharing ovs builds between tests, support selecting
the ovn/ovn build directory from the outside.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>