Make sure we register the notifier before the initial dump
If we fetch the databases' content before setting the notifier, there is
a chance of lost events that are produced in between. To avoid that, we
register the notifier first and then we do the initial database dump.

In order to make sure the initialization is served before the notifier
callbacks, we lock the cachemutex beforehand.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>