Close rpc2 client connection during txn error
This is a workaround due to a bug in ovsdb-server where it may silently
drop monitors:
https://bugzilla.redhat.com/show_bug.cgi?id=2000375

Related to ovnkube bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1998614

tl;dr ovsdb-server may drop the monitor to go-ovn silently which will
cause the go-ovn cache to go stale. Subsequent retry txns for the
resource will result in errors from ovsdb-server (since the resource may
have been created in ovsdb, but we are not getting updates) threfore a
txn error is encountered force reconnect to rebuild cache and get a new
monitor.

Co-authored-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>

Signed-off-by: Tim Rozet <trozet@redhat.com>